fix: add helpers to compatibility.js
